Fail on restore failures by default, instead of warn
WarnAndContinue seemed like a nicer default, but it ended up causing unnecessary 
manual log parsing to find those warnings in a the (potentially big) list of warnings 
for a project, when failing to restore will typically cause a build to fail instead. 

Users can still default to that previous value, but it's better to default to ErrorAndContinue 
instead.